+++ src/test/java/com/mageddo/dnsproxyserver/templates/MessageTemplates.java
@@ -8,6 +8,9 @@ import org.xbill.DNS.Record;
 import org.xbill.DNS.Type;
 
 public class MessageTemplates {
+
+  public static final String ACME_HOSTNAME = "acme.com";
+
   @SneakyThrows
   public static Message buildAQuestionFor(String hostname) {
     final var r = Record.newRecord(Name.fromString(hostname), Type.A, DClass.IN, 0);
@@ -15,7 +18,8 @@ public class MessageTemplates {
   }
 
   public static Message acmeAQuery() {
-    return buildAQuestionFor("acme.com.");
+    return buildAQuestionFor(ACME_HOSTNAME + ".");
   }
+
 
 }